+++ src/Symfony/Component/DependencyInjection/ContainerAwareTrait.php
@@ -25,6 +25,10 @@ trait ContainerAwareTrait
 
     public function setContainer(ContainerInterface $container = null)
     {
+        if (1 > \func_num_args()) {
+            trigger_deprecation('symfony/dependency-injection', '6.2', 'Calling "%s::%s()" without any arguments is deprecated. Please explicitly pass null if you want to unset the container.', static::class, __FUNCTION__);
+        }
+
         $this->container = $container;
     }
 }