+++ src/Symfony/Component/Security/Core/Authentication/Token/Storage/TokenStorage.php
@@ -39,6 +39,10 @@ class TokenStorage implements TokenStorageInterface, ResetInterface
 
     public function setToken(TokenInterface $token = null)
     {
+        if (1 > \func_num_args()) {
+            trigger_deprecation('symfony/security-core', '6.2', 'Calling "%s()" without any arguments is deprecated. Please explicitly pass null if you want to unset the token.', __METHOD__);
+        }
+
         if ($token) {
             // ensure any initializer is called
             $this->getToken();